[enh #35] 便利マクロなどの追加
-CommonフォルダをUtilityフォルダに名前変更
-テンプレート関数をまとめたヘッダと解放テンプレート関数追加
-マクロの追加
-一部ソース修正
