+++ LibraryProject/SharokuLibrary/sl/Library/Utility/slDefine.h
@@ -1,0 +1,31 @@
+﻿//==================================================================================================================================//
+//!< @file		slDefine.h
+//!< @brief		マクロ関連をまとめているヘッダファイル
+//!< @author	T.Haga
+//!< @data		作成日時：2017/08/30
+//==================================================================================================================================//
+
+#ifndef SL_DEFINE_H
+#define SL_DEFINE_H
+
+#include "../../slBuild.h"
+
+/** インターフェイスマクロ */
+#ifndef	interface
+	#define interface struct
+#endif	// #ifndef	interface 
+
+/** falseかどうかのチェックマクロ */
+#define RESULT_FALSE(result) ((static_cast<bool>(result)) == false)
+
+/** コピー禁止マクロ */
+#define SL_DISALLOW_COPY_AND_ASSIGN(...)			\
+__VA_ARGS__(const __VA_ARGS__&) = delete;			\
+void operator=(const __VA_ARGS__&) = delete;
+
+
+#endif	// SL_DEFINE_H
+
+//==================================================================================================================================//
+// END OF FILE
+//==================================================================================================================================//
