+++ LibraryProject/SharokuLibrary/sl/Library/Utility/slUtility.h
@@ -1,0 +1,28 @@
+﻿//==================================================================================================================================//
+//!< @file		slUtility.h
+//!< @brief		Utilityフォルダ内のヘッダファイルを集めたヘッダ
+//!< @author	T.Haga
+//!< @data		作成日時：2017/09/10
+//==================================================================================================================================//
+
+#ifndef SL_UTILITY_H
+#define SL_UTILITY_H
+
+/** マクロを関連をまとめたヘッダ */
+#include "slDefine.h"
+
+/** Rect構造体(矩形)のヘッダ */
+#include "slRect.h"
+
+/** テンプレート関数をまとめたヘッダ */
+#include "slTemplateFunction.h"
+
+/** 型定義をまとめたヘッダ */
+#include "slTypes.h"
+
+#endif	// SL_UTILITY_H
+
+//==================================================================================================================================//
+// END OF FILE
+//==================================================================================================================================//
+
